+++ pkg/azure/infrastructure.go
@@ -6,9 +6,5 @@ var infrastructure = map[string][]string{
 		"2",
 		"3",
 	},
-	"germanywestcentral": {
-		"1",
-		"2",
-		"3",
-	},
+	"germanywestcentral": {},
 }